clean up my repository

+++ test4dalloc.cpp
@@ -2,7 +2,7 @@
 
 int main(void){
 
-  long double **** my4dptr = allocate4d(3,10,10,10);
+  double **** my4dptr = allocate4d(3,10,10,10);
 
   std::cout << my4dptr[2][3][4][5] << std::endl;
   my4dptr[2][3][4][5] = 3.0;
